server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
psycopg[binary]==3.2.3
pgvector==0.3.6
sqlalchemy[asyncio]==2.0.36
pydantic-settings==2.1.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.3
scipy==1.11.4
hdbscan==0.8.33
umap-learn==0.5.5
pytest==7.4.3
pyyaml==6.0.1
pyarrow==14.0.1
httpx==0.27.0
sentence-transformers>=3.0.0
torch>=2.0.0# ChromaDB 검색 관련 패키지
anthropic>=0.34.0
langchain-upstage>=0.1.0
langchain-chroma>=0.1.0
chromadb>=0.4.0
